+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Changelog
 Added
 ~~~~~
 
+- Added `torch.PyTorch`: Apply a PyTorch module. (#95)
+
 - Added `batch.BatchPipeline`: Combine consecutive objects into a batch. (#92)
 
 - Added `mjpeg_streamer.MJPEGStreamer`: Stream images via HTTP (e.g. to the Browser). (#75)
@@ -28,6 +30,10 @@ Added
 
 Changed
 ~~~~~~~
+
+- EcotaxaReader: Return EcotaxaObject. (#102)
+
+- EcotaxaWriter: Allow Variables for `archive_fn`. (#100)
 
 - Use `UnavailableObject` instead of `import_optional_dependency`.